Add option for skipping expiration certain entries
This change adds functionality for skipping the expiration of certain
entries based on a user-supplied function. If the function returns true,
the the entry will not be expired. By default, no entries are skipped.

This functionality is being added so cached databases in couch_server
will only expire if the database process has terminated.

BugzID: 19132